General Cleanup of Frontend
