+++ RetireSimple.Frontend/package.json
@@ -56,7 +56,10 @@
 			],
 			"indent": [
 				"error",
-				"tab"
+				"tab",
+				{
+					"SwitchCase": 1
+				}
 			],
 			"brace-style": [
 				"error",
