+++ RetireSimple.Frontend/src/api/ApiCommon.ts
@@ -1,4 +1,4 @@
-import {Portfolio} from "../data/Interfaces";
+import {Portfolio} from "../Interfaces";
 
 export const API_BASE_URL = process.env.REACT_APP_API_URL;
