+++ RetireSimple.Frontend/src/forms/InvestmentDataForm.tsx
@@ -1,7 +1,7 @@
 import {Box, Grid} from '@mui/material';
 import React from 'react';
 import {useFormContext, useWatch} from 'react-hook-form';
-import {FormSelectField, FormTextField} from './Inputs';
+import {FormSelectField, FormTextField} from '../components/InputComponents';
 import {BondForm} from './investment/BondForm';
 import {StockForm} from './investment/StockForm';
 
@@ -18,9 +18,8 @@ export const InvestmentDataForm = (props: InvestmentDataFormProps) => {
 		name: 'investmentType',
 		control: formContext.control,
 		defaultValue: 'StockInvestment',
-
 	});
-	const { errors } = formContext.formState;
+	const {errors} = formContext.formState;
 
 	//==============================================
 	//Field definitions (To reduce indent depth)
@@ -31,7 +30,8 @@ export const InvestmentDataForm = (props: InvestmentDataFormProps) => {
 			label='Name'
 			control={formContext.control}
 			errorField={errors.investmentName}
-		/>);
+		/>
+	);
 
 	const investmentTypeField = (
 		<FormSelectField
@@ -45,43 +45,56 @@ export const InvestmentDataForm = (props: InvestmentDataFormProps) => {
 				{value: 'BondInvestment', label: 'Bond'},
 			]}
 			disable={props.disableTypeSelect ?? false}
-		/>);
-
-
+		/>
+	);
 
 	const investmentTypeSubform = React.useMemo(() => {
 		switch (investmentType) {
-		case 'StockInvestment':
-			return <StockForm analysisTypeField={
-				<FormSelectField
-					name='analysisType'
-					label='Analysis Type'
-					control={formContext.control}
-					errorField={errors.analysisType}
-					options={[
-						{value: 'MonteCarlo_NormalDist', label: 'Monte Carlo (Normal Dist)'},
-						{value: 'MonteCarlo_LogNormalDist', label: 'Monte Carlo (Log Normal Dist)'},
-					]}
-					defaultOption=''
-					disable={false}
-				/>
-			} />;
-		case 'BondInvestment':
-			return <BondForm analyisisTypeField={
-				<FormSelectField
-					name='analysisType'
-					label='Analysis Type'
-					control={formContext.control}
-					errorField={errors.analysisType}
-					options={[
-						{ value: 'bondValuationAnalysis', label: 'Bond Valuation' },
-					]}
-					defaultOption=''
-					disable={false}
-				/>
-			} />;
-		default:
-			return <div>Unknown investment type</div>;
+			case 'StockInvestment':
+				return (
+					<StockForm
+						analysisTypeField={
+							<FormSelectField
+								name='analysisType'
+								label='Analysis Type'
+								control={formContext.control}
+								errorField={errors.analysisType}
+								options={[
+									{
+										value: 'MonteCarlo_NormalDist',
+										label: 'Monte Carlo (Normal Dist)',
+									},
+									{
+										value: 'MonteCarlo_LogNormalDist',
+										label: 'Monte Carlo (Log Normal Dist)',
+									},
+								]}
+								defaultOption=''
+								disable={false}
+							/>
+						}
+					/>
+				);
+			case 'BondInvestment':
+				return (
+					<BondForm
+						analyisisTypeField={
+							<FormSelectField
+								name='analysisType'
+								label='Analysis Type'
+								control={formContext.control}
+								errorField={errors.analysisType}
+								options={[
+									{value: 'bondValuationAnalysis', label: 'Bond Valuation'},
+								]}
+								defaultOption=''
+								disable={false}
+							/>
+						}
+					/>
+				);
+			default:
+				return <div>Unknown investment type</div>;
 		}
 	}, [investmentType, formContext.control, errors]);
 
@@ -89,11 +102,16 @@ export const InvestmentDataForm = (props: InvestmentDataFormProps) => {
 		<>
 			<Box>
 				<Grid container spacing={2}>
-					<Grid item xs={4}>{investmentNameField}</Grid>
-					<Grid item xs={4}>{investmentTypeField}</Grid>
+					<Grid item xs={4}>
+						{investmentNameField}
+					</Grid>
+					<Grid item xs={4}>
+						{investmentTypeField}
+					</Grid>
 				</Grid>
 				{investmentTypeSubform}
 			</Box>
 			{props.children}
-		</>);
+		</>
+	);
 };
