+++ RetireSimple.Frontend/src/forms/analysis/BondValuationAnalysisForm.tsx
@@ -1,7 +1,7 @@
 import {Grid, Typography} from '@mui/material';
 import React from 'react';
 import {useFormContext} from 'react-hook-form';
-import {FormSelectField, FormTextField} from '../Inputs';
+import {FormSelectField, FormTextField} from '../../components/InputComponents';
 
 export const BondValuationAnalysisForm = () => {
 	const formContext = useFormContext();
@@ -16,27 +16,34 @@ export const BondValuationAnalysisForm = () => {
 			label='Analysis Length (Months)'
 			control={formContext.control}
 			errorField={errors.analysis_analysisLength}
-		/>);
+		/>
+	);
 	const analysisIsAnnualField = (
 		<FormSelectField
 			name='analysis_isAnnual'
 			label='Analysis is Annual'
 			control={formContext.control}
 			options={[
-				{ value: 'true', label: 'Annual' },
-				{ value: 'false', label: 'Semi Annual' },
+				{value: 'true', label: 'Annual'},
+				{value: 'false', label: 'Semi Annual'},
 			]}
 			defaultOption={'true'}
-			disable={false} 
-			errorField={errors.analysis_isAnnual} />);
+			disable={false}
+			errorField={errors.analysis_isAnnual}
+		/>
+	);
 
 	return (
 		<Grid container spacing={2}>
 			<Grid item xs={12}>
 				<Typography variant='subtitle2'>Bond Valuation Parameters</Typography>
 			</Grid>
-			<Grid item xs={4}>{analysisLengthField}</Grid>
-			<Grid item xs={4}>{analysisIsAnnualField}</Grid>
+			<Grid item xs={4}>
+				{analysisLengthField}
+			</Grid>
+			<Grid item xs={4}>
+				{analysisIsAnnualField}
+			</Grid>
 		</Grid>
 	);
 };
